+++ scusocial/lib/services/firestore_service.dart
@@ -1,0 +1,91 @@
+import 'package:cloud_firestore/cloud_firestore.dart';
+import 'package:flutter/material.dart';
+
+class FirestoreService {
+  final FirebaseFirestore _firestore = FirebaseFirestore.instance;
+
+  Stream<List<Map<String, dynamic>>> getEventsStream() {
+    return _firestore.collection('events').snapshots().map((snapshot) =>
+        snapshot.docs.map((doc) => {...doc.data(), 'id': doc.id}).toList());
+  }
+
+  Future<void> createEvent(BuildContext context, String userId) async {
+    final nameController = TextEditingController();
+    final descriptionController = TextEditingController();
+    final locationController = TextEditingController();
+    DateTime? selectedDate;
+    TimeOfDay? selectedTime;
+
+    showDialog(
+      context: context,
+      builder: (context) {
+        return AlertDialog(
+          title: Text('Create Event'),
+          content: Column(
+            mainAxisSize: MainAxisSize.min,
+            children: [
+              TextField(
+                controller: nameController,
+                decoration: InputDecoration(labelText: 'Event Name'),
+              ),
+              TextField(
+                controller: descriptionController,
+                decoration: InputDecoration(labelText: 'Description'),
+              ),
+              TextField(
+                controller: locationController,
+                decoration: InputDecoration(labelText: 'Location'),
+              ),
+              ElevatedButton(
+                onPressed: () async {
+                  selectedDate = await showDatePicker(
+                    context: context,
+                    initialDate: DateTime.now(),
+                    firstDate: DateTime.now(),
+                    lastDate: DateTime(2100),
+                  );
+                },
+                child: Text('Select Date'),
+              ),
+              ElevatedButton(
+                onPressed: () async {
+                  selectedTime = await showTimePicker(
+                    context: context,
+                    initialTime: TimeOfDay.now(),
+                  );
+                },
+                child: Text('Select Time'),
+              ),
+            ],
+          ),
+          actions: [
+            ElevatedButton(
+              onPressed: () async {
+                if (nameController.text.isNotEmpty &&
+                    selectedDate != null &&
+                    selectedTime != null &&
+                    descriptionController.text.isNotEmpty &&
+                    locationController.text.isNotEmpty) {
+                  final eventTime = selectedTime!.format(context);
+                  await _firestore.collection('events').add({
+                    'name': nameController.text,
+                    'description': descriptionController.text,
+                    'location': locationController.text,
+                    'date': selectedDate,
+                    'time': eventTime,
+                    'creatorId': userId,
+                    'accepted': [],
+                  });
+                  Navigator.pop(context);
+                } else {
+                  print('Please fill all fields.');
+                }
+              },
+              child: Text('Create'),
+            ),
+          ],
+        );
+      },
+    );
+  }
+}
